somethings

+++ JonnyGota/JonnyGota/Objects/JAGHud.m
@@ -78,7 +78,7 @@
     _tempo.position=CGPointMake(size.width*0.5, size.height*0.94);
     
     _saude=[[SKLabelNode alloc]initWithFontNamed:fonte];
-    _saude.text=[NSString stringWithFormat:@"saúde: %d", _gota.vida];
+    _saude.text=[NSString stringWithFormat:@"saúde: %ld", (long)_gota.vida];
     _saude.fontSize = size.height*0.05;
     _saude.position=CGPointMake(size.width*0.73, size.height*0.94);
